+++ packages/firebase_admob/lib/src/ad_instance_manager.dart
@@ -4,7 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:quiver/collection.dart';
 
-import 'ads.dart';
+import 'ad.dart';
 
 // ignore_for_file: public_member_api_docs
